+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.3.0] - 2024-06-14
 ### Fixed
 - Fixed issue with wrong argument name for `nakama.rpc_func`
 - Updated several of the socket messages so that they no longer incorrectly wait for a response from the server.